More Arduino sync stuff
Arduino working to turn on and off LED as SLM drive signal changes.

+++ Arduino/slmSyncTest_serialPrint/slmSyncTest_serialPrint.ino
@@ -38,7 +38,7 @@ void setup()
 void loop() {
    
   // read from the sensor:  
-  readings[index] = (byte) digitalRead( slmInput );
+  readings[index] = digitalRead( slmInput );
   readingTimes[index] = (int) micros();
       
   // advance to the next position in the array:
